backend/requirements.txt
Django==5.2.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
setuptools>=70.0.0
psycopg[binary]==3.2.13
Pillow==10.4.0
python-decouple==3.8
celery==5.3.6
redis==5.0.8
django-filter==24.3
drf-spectacular==0.27.2
gunicorn==22.0.0
whitenoise==6.7.0
qrcode==7.4.2
svgwrite==1.4.3
reportlab==4.2.0
python-pptx==0.6.23
openpyxl==3.1.5
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.0
faker==30.1.0
django-ratelimit==4.1.0
